Fix integer truncation in calculate_reduced_time with int time arrays
reduced_time = np.zeros_like(time) inherited the input dtype, so an integer
time array (e.g. np.arange(0, 10)) made each trapezoidal accumulation
truncate to int, silently returning wrong results. Allocating the
accumulator as float fixes this; verified against the float-input path in
the regression test.

+++ aerosandbox/library/aerodynamics/unsteady.py
@@ -51,7 +51,7 @@ def calculate_reduced_time(
         assert np.size(velocity) == np.size(time), (
             "The velocity history and time must have the same length"
         )
-        reduced_time = np.zeros_like(time)
+        reduced_time = np.zeros_like(time, dtype=float)
         for i in range(len(time) - 1):
             reduced_time[i + 1] = reduced_time[i] + (
                 velocity[i + 1] + velocity[i]
